requirements.txt
torch==2.4.0
torchvision==0.19.0
transformers==4.44.2
git+https://github.com/openai/CLIP.git
sentencepiece==0.1.99
beautifulsoup4
matplotlib
nltk
numpy
pandas
Pillow
scikit-learn
seaborn
tqdm
gdown
Madgrad
openpyxl
multilingual-clip torch